requirements.txt
Django==2.1.2
djangorestframework==3.9.0
drf-nested-routers==0.91
psycopg2==2.7.5
pytz==2018.5
